07-Data-Engineering/03-Train-at-scale/03-Train-taxiFare-on-gcp/requirements.txt
pandas
google-cloud-storage
scikit-learn
numpy
joblib
gcsfs
six==1.13.0
pip>=9
setuptools>=26
wheel>=0.29
pytest
coverage
flake8
black
yapf
python-gitlab
twine
memoized_property
termcolor
category_encoders
pygeohash==1.2.0
mlflow==1.12.1
xgboost==1.2.1
psutil==5.7.3
joblib==0.17.0
protobuf==3.14.0
